Create workspace, move sample macro for testing there. (#7)
Create a workspace of packages in order to facilitate testing (and
not be forced to make lib of `proc_macro = true` type. A new package
(samplemacros) currently containts the daft `rename` macro which
we test in `test_sanity`. In addition, update `rust.yml` in
workflows to use `test --all` (test *all* packages).